requirements.txt
opencv-python
numpy
pillow
fpdf